fixing warnings

+++ inc/ui/webhooks-table.php
@@ -16,20 +16,22 @@ $subscriptions = get_strike_subscriptions();
 </tr>
 <?php
 
-foreach ( $subscriptions as $sub ){
-  if ($sub['enabled']=='1'){ $check = 'checked'; } else { $check = ''; }
+if ( isset($subscriptions[0]['enabled']) ){
+  
+  foreach ( $subscriptions as $sub ){
+    if ($sub['enabled']=='1'){ $check = 'checked'; } else { $check = ''; }
 
-  echo '<tr>
-    <td>'.$sub['id'].'</td><td>'.$sub['webhookUrl'].'</td>
-    <td>'.$sub['webhookVersion'].'</td><td>'.$sub['enabled'].'</td>
-    <td>'.$sub['created'].'</td><td>'.json_encode($sub['eventTypes']).'</td>
-    <td>
-    <a href="'.strtok($_SERVER["REQUEST_URI"], '?').'?deleteSubscription='.$sub['id'].'">X
-    </a>
-    </td>
-    </tr>';    
+    echo '<tr>
+      <td>'.$sub['id'].'</td><td>'.$sub['webhookUrl'].'</td>
+      <td>'.$sub['webhookVersion'].'</td><td>'.$sub['enabled'].'</td>
+      <td>'.$sub['created'].'</td><td>'.json_encode($sub['eventTypes']).'</td>
+      <td>
+      <a href="'.strtok($_SERVER["REQUEST_URI"], '?').'?deleteSubscription='.$sub['id'].'">X
+      </a>
+      </td>
+      </tr>';    
+  }
 }
-
 ?>
 
 </table>
